requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
streamlit==1.30.0
openai==1.10.0
requests==2.31.0
pypdf==3.17.4
faiss-cpu==1.7.4
numpy==1.26.3
sentence-transformers==2.3.1
python-dotenv==1.0.0
pydantic==2.5.3
